// src/app/models/setting.ts
export class Setting {
  id: number;
  w_color: string;
  w_group_by_category: number;
  w_show_any_employee: number;
  w_step_display: number;
  w_step_search: number;
  w_let_check_steps: number;
  w_steps_service?: string[];
  w_steps_employee?: string[];
  w_steps_g?: string[];
  w_contact_step: string;
  w_to_group_category: number;
}